+++ api/prisma/seed-helpers/translation-factory.ts
@@ -1,4 +1,5 @@
-import { LanguagesEnum, Prisma } from '@prisma/client';
+import { LanguagesEnum, Prisma, PrismaClient } from '@prisma/client';
+import * as lodash from 'lodash';
 
 const translations = (
   jurisdiction?: {
@@ -653,6 +654,34 @@ const translations = (
       };
   }
 };
+
+export async function upsertTranslation(
+  prisma: PrismaClient,
+  data: Prisma.TranslationsCreateInput,
+): Promise<void> {
+  const jurisdictionId = data.jurisdictions?.connect?.id ?? null;
+  const { language } = data;
+
+  const existing = await prisma.translations.findFirst({
+    where: { language, jurisdictionId },
+  });
+
+  if (!existing) {
+    await prisma.translations.create({ data });
+    return;
+  }
+
+  const merged = lodash.merge(
+    {},
+    existing.translations as Record<string, unknown>,
+    data.translations as Record<string, unknown>,
+  );
+
+  await prisma.translations.update({
+    where: { id: existing.id },
+    data: { translations: merged as Prisma.InputJsonValue },
+  });
+}
 
 export const translationFactory = (optionalParams?: {
   jurisdiction?: {
